src/main/java/com/idea5/four_cut_photos_map/domain/member/service/MemberReadServiceImpl.java
package com.idea5.four_cut_photos_map.domain.member.service;


import com.idea5.four_cut_photos_map.domain.member.entity.Member;
import com.idea5.four_cut_photos_map.domain.member.repository.MemberRepository;
import com.idea5.four_cut_photos_map.global.error.ErrorCode;
import com.idea5.four_cut_photos_map.global.error.exception.BusinessException;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class MemberReadServiceImpl implements MemberReadService {
    private final MemberRepository memberRepository;

    @Override
    public Member getMemberWithThrow(Long memberId) {
        return memberRepository.findById(memberId)
                .orElseThrow(() -> new BusinessException(ErrorCode.MEMBER_NOT_FOUND));
    }
}
